svm: merge FieldAccessorImpl and UnsafeFieldAccessorImpl
